消息的修改，添加录音部分（还待完善）

+++ CWWeChat/MainClass/Message/Model/Message/CWVoiceMessageContent.swift
@@ -11,7 +11,7 @@ import UIKit
 class CWVoiceMessageContent: CWMessageContent {
 
     /// 录音长度
-    var voiceLength: Int?
+    var voiceLength: Float?
     /// 网络路径
     var voiceURL: String?
     /// 本地路径
